acorn is now called on the unmodified source file to report errors.
This commit fixes an issue with the error message reporting syntax errors.

For example, with the following myModule.js file (containing 3 lines):

    // The following line intentionally contains an error
    myError.
    // there should be something after the dot

The reported error contained:

   Error: Unexpected token in 'myModule.js' (line 5, column 0):
   })
   ^-
   failed to load definition of module 'myModule.js'
   failed to preload 'myModule.js'

The error is line 5 whereas the source file only contains 3 lines.
That's because noder wraps the code in a function (making the file
larger than 3 lines).

This commit fixes this issue by making sure the code passed to acorn
to validate the syntax is only the source code, without the wrapping
function.

close #10